+++ packages/js/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @convergence-rfq/sdk
 
+## 4.0.12
+
+### Patch Changes
+
+- change type of rfq pda timestamp buffer
+
 ## 4.0.11
 
 ### Patch Changes